UP3 Patch Pack 304 WIP

+++ Storebror/UP3 Patch Pack 3xx/src/com/maddox/il2/objects/air/CockpitPilot.java
@@ -17,46 +17,40 @@ import com.maddox.il2.net.NetMissionTrack;
 import com.maddox.rts.Property;
 import com.maddox.rts.Time;
 
-public class CockpitPilot extends Cockpit
-{
-	private String[]			hotKeyEnvs		= { "pilot", "move" };
-	protected float				stepAzimut		= 45.0F;
-	protected float				stepTangage		= 30.0F;
-	protected float				minMaxAzimut	= 145.0F;
-	protected float				maxTangage		= 90.0F;
-	protected float				minTangage		= -60.0F;
-	protected Point3d			cameraCenter	= new Point3d();
-	private Point3d				cameraAim;
-	private Point3d				cameraUp;
-	private boolean				bBeaconKeysEnabled;
-	private float				saveZN;
-	protected float				normZN;
-	protected float				gsZN;
-	private double				pictBall		= 0.0;
-	private long				oldBallTime		= 0L;
+public class CockpitPilot extends Cockpit {
+	private String[] hotKeyEnvs = { "pilot", "move" };
+	protected float stepAzimut = 45.0F;
+	protected float stepTangage = 30.0F;
+	protected float minMaxAzimut = 145.0F;
+	protected float maxTangage = 90.0F;
+	protected float minTangage = -60.0F;
+	protected Point3d cameraCenter = new Point3d();
+	private Point3d cameraAim;
+	private Point3d cameraUp;
+	private boolean bBeaconKeysEnabled;
+	private float saveZN;
+	protected float normZN;
+	protected float gsZN;
+	private double pictBall = 0.0;
+	private long oldBallTime = 0L;
 
-	class Interpolater extends InterpolateRef
-	{
-		public boolean tick()
-		{
+	class Interpolater extends InterpolateRef {
+		public boolean tick() {
 			if (isPadlock())
 				HookPilot.current.checkPadlockState();
 			return true;
 		}
 	}
 
-	public boolean isEnableHotKeysOnOutsideView()
-	{
+	public boolean isEnableHotKeysOnOutsideView() {
 		return true;
 	}
 
-	public String[] getHotKeyEnvs()
-	{
+	public String[] getHotKeyEnvs() {
 		return hotKeyEnvs;
 	}
 
-	private float ZNear(float f)
-	{
+	private float ZNear(float f) {
 		if (f < 0.0F)
 			return -1.0F;
 		Camera3D camera3d = (Camera3D) Actor.getByName("camera");
@@ -65,29 +59,24 @@ public class CockpitPilot extends Cockpit
 		return f_0_;
 	}
 
-	protected boolean doFocusEnter()
-	{
+	protected boolean doFocusEnter() {
 		HookPilot hookpilot = HookPilot.current;
 		Aircraft aircraft = aircraft();
 		Main3D main3d = Main3D.cur3D();
 		hookpilot.setCenter(cameraCenter);
 		hookpilot.setAim(cameraAim);
 		hookpilot.setUp(cameraUp);
-		if (!NetMissionTrack.isPlaying() || NetMissionTrack.playingOriginalVersion() > 101)
-		{
+		if (!NetMissionTrack.isPlaying() || NetMissionTrack.playingOriginalVersion() > 101) {
 			hookpilot.setSteps(stepAzimut, stepTangage);
 			hookpilot.setMinMax(minMaxAzimut, minTangage, maxTangage);
-		}
-		else
-		{
+		} else {
 			hookpilot.setSteps(45.0F, 30.0F);
 			hookpilot.setMinMax(135.0F, -60.0F, 90.0F);
 		}
 		hookpilot.reset();
 		hookpilot.use(true);
 		aircraft.setAcoustics(acoustics);
-		if (acoustics != null)
-		{
+		if (acoustics != null) {
 			aircraft.enableDoorSnd(true);
 			if (acoustics.getEnvNum() == 2)
 				aircraft.setDoorSnd(1.0F);
@@ -109,8 +98,7 @@ public class CockpitPilot extends Cockpit
 		return true;
 	}
 
-	protected void doFocusLeave()
-	{
+	protected void doFocusLeave() {
 		saveZN = ZNear(saveZN);
 		((AircraftLH) aircraft()).bWantBeaconKeys = bBeaconKeysEnabled;
 		HookPilot hookpilot = HookPilot.current;
@@ -131,74 +119,61 @@ public class CockpitPilot extends Cockpit
 		aircraft().setMotorPos(null);
 	}
 
-	public boolean existPadlock()
-	{
+	public boolean existPadlock() {
 		return true;
 	}
 
-	public boolean isPadlock()
-	{
+	public boolean isPadlock() {
 		if (!isFocused())
 			return false;
 		HookPilot hookpilot = HookPilot.current;
 		return hookpilot.isPadlock();
 	}
 
-	public Actor getPadlockEnemy()
-	{
+	public Actor getPadlockEnemy() {
 		if (!isFocused())
 			return null;
 		HookPilot hookpilot = HookPilot.current;
 		return hookpilot.isPadlock() ? hookpilot.getEnemy() : null;
 	}
 
-	public boolean startPadlock(Actor actor)
-	{
+	public boolean startPadlock(Actor actor) {
 		if (!isFocused())
 			return false;
 		HookPilot hookpilot = HookPilot.current;
 		return hookpilot.startPadlock(actor);
 	}
 
-	public void stopPadlock()
-	{
-		if (isFocused())
-		{
+	public void stopPadlock() {
+		if (isFocused()) {
 			HookPilot hookpilot = HookPilot.current;
 			hookpilot.stopPadlock();
 		}
 	}
 
-	public void endPadlock()
-	{
-		if (isFocused())
-		{
+	public void endPadlock() {
+		if (isFocused()) {
 			HookPilot hookpilot = HookPilot.current;
 			hookpilot.endPadlock();
 		}
 	}
 
-	public void setPadlockForward(boolean bool)
-	{
-		if (isFocused())
-		{
+	public void setPadlockForward(boolean bool) {
+		if (isFocused()) {
 			HookPilot hookpilot = HookPilot.current;
 			hookpilot.setForward(bool);
 		}
 	}
 
-	public boolean isToggleAim()
-	{
+	public boolean isToggleAim() {
 		if (!isFocused())
 			return false;
 		HookPilot hookpilot = HookPilot.current;
 		return hookpilot.isAim();
 	}
 
-	public void doToggleAim(boolean bool)
-	{
-		if (isFocused())
-		{
+	public void doToggleAim(boolean bool) {
+		if (isFocused()) {
 			HookPilot hookpilot = HookPilot.current;
 			hookpilot.doAim(bool);
 			if (bool)
@@ -208,53 +183,43 @@ public class CockpitPilot extends Cockpit
 		}
 	}
 
-	public boolean isToggleUp()
-	{
+	public boolean isToggleUp() {
 		if (!isFocused())
 			return false;
 		HookPilot hookpilot = HookPilot.current;
 		return hookpilot.isUp();
 	}
 
-	public void doToggleUp(boolean bool)
-	{
-		if (isFocused() && (!bool || cameraUp != null))
-		{
+	public void doToggleUp(boolean bool) {
+		if (isFocused() && (!bool || cameraUp != null)) {
 			HookPilot hookpilot = HookPilot.current;
 			hookpilot.doUp(bool);
 		}
 	}
 
-	public CockpitPilot(String string, String string_1_)
-	{
+	public CockpitPilot(String string, String string_1_) {
 		super(string, string_1_);
 		HookNamed hooknamed = new HookNamed(mesh, "CAMERA");
 		Loc loc = new Loc();
 		hooknamed.computePos(this, pos.getAbs(), loc);
 		loc.get(cameraCenter);
-		try
-		{
+		try {
 			hooknamed = new HookNamed(mesh, "CAMERAAIM");
 			loc.set(0.0, 0.0, 0.0, 0.0F, 0.0F, 0.0F);
 			hooknamed.computePos(this, pos.getAbs(), loc);
 			cameraAim = new Point3d();
 			loc.get(cameraAim);
-		}
-		catch (Exception exception)
-		{
+		} catch (Exception exception) {
 			System.out.println(exception.getMessage());
 			exception.printStackTrace();
 		}
-		try
-		{
+		try {
 			hooknamed = new HookNamed(mesh, "CAMERAUP");
 			loc.set(0.0, 0.0, 0.0, 0.0F, 0.0F, 0.0F);
 			hooknamed.computePos(this, pos.getAbs(), loc);
 			cameraUp = new Point3d();
 			loc.get(cameraUp);
-		}
-		catch (Exception exception)
-		{
+		} catch (Exception exception) {
 			/* empty */
 		}
 		pos.setBase(aircraft(), new Cockpit.HookOnlyOrient(), false);
@@ -263,37 +228,32 @@ public class CockpitPilot extends Cockpit
 			HookPilot.current.doUp(false);
 		normZN = Property.floatValue(this.getClass(), "normZN", -1.0F);
 		gsZN = Property.floatValue(this.getClass(), "gsZN", -1.0F);
-		
-		//TODO: Added by |ZUTI|: reset multi crew bomber position
-		//-----------------------------------------------------------
-		if( string.indexOf("Bombardier") > -1 )
-		{
-			pos.setAbs(new Point3d(0,0,0), new Orient(0,0,0));
-			//System.out.println("BOMBARDIER COCKPIT POSITION RESET!");
+
+		// TODO: Added by |ZUTI|: reset multi crew bomber position
+		// -----------------------------------------------------------
+		if (string.indexOf("Bombardier") > -1) {
+			pos.setAbs(new Point3d(0, 0, 0), new Orient(0, 0, 0));
+			// System.out.println("BOMBARDIER COCKPIT POSITION RESET!");
 		}
-		//-----------------------------------------------------------
+		// -----------------------------------------------------------
 	}
 
-	protected float getBall(double d)
-	{
+	protected float getBall(double d) {
 		double d_2_ = 0.0;
 		long l = Time.current();
 		long l_3_ = l - oldBallTime;
 		oldBallTime = l;
 		if (l_3_ > 200L)
 			l_3_ = 200L;
-		double d_4_ = 3.8E-4 * (double) l_3_;
-		if (-fm.getBallAccel().z > 0.0010)
-		{
+		double d_4_ = 3.8E-4 * l_3_;
+		if (-fm.getBallAccel().z > 0.0010) {
 			d_2_ = Math.toDegrees(Math.atan2(fm.getBallAccel().y, -fm.getBallAccel().z));
 			if (d_2_ > 20.0)
 				d_2_ = 20.0;
 			else if (d_2_ < -20.0)
 				d_2_ = -20.0;
 			pictBall = (1.0 - d_4_) * pictBall + d_4_ * d_2_;
-		}
-		else
-		{
+		} else {
 			if (pictBall > 0.0)
 				d_2_ = 20.0;
 			else
@@ -307,8 +267,7 @@ public class CockpitPilot extends Cockpit
 		return (float) pictBall;
 	}
 
-	static
-	{
+	static {
 		Property.set(CockpitPilot.class, "astatePilotIndx", 0);
 	}
 }
